Rename module test.on_yubikey.cli_piv.test_cli_piv to [...].util
